api/requirements.txt
discord.py
flask
google-auth
google-cloud-firestore
google-cloud-storage
Pillow
requests
